--- face_recognition/face_recognition.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.1.32228.430
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "face_recognition", "face_recognition\face_recognition.vcxproj", "{6FE84039-179C-420A-9205-435A576FB054}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6FE84039-179C-420A-9205-435A576FB054}.Debug|x64.ActiveCfg = Debug|x64
-		{6FE84039-179C-420A-9205-435A576FB054}.Debug|x64.Build.0 = Debug|x64
-		{6FE84039-179C-420A-9205-435A576FB054}.Debug|x86.ActiveCfg = Debug|Win32
-		{6FE84039-179C-420A-9205-435A576FB054}.Debug|x86.Build.0 = Debug|Win32
-		{6FE84039-179C-420A-9205-435A576FB054}.Release|x64.ActiveCfg = Release|x64
-		{6FE84039-179C-420A-9205-435A576FB054}.Release|x64.Build.0 = Release|x64
-		{6FE84039-179C-420A-9205-435A576FB054}.Release|x86.ActiveCfg = Release|Win32
-		{6FE84039-179C-420A-9205-435A576FB054}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {1F4206A0-863B-4B81-AEEC-A7D9C295F820}
-	EndGlobalSection
-EndGlobal
