--- image_contour/image_contour.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.32228.343
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "image_contour", "image_contour.vcxproj", "{6C82FDF0-ABCA-40FE-BE18-E900BC35F128}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6C82FDF0-ABCA-40FE-BE18-E900BC35F128}.Debug|x64.ActiveCfg = Debug|x64
-		{6C82FDF0-ABCA-40FE-BE18-E900BC35F128}.Debug|x64.Build.0 = Debug|x64
-		{6C82FDF0-ABCA-40FE-BE18-E900BC35F128}.Debug|x86.ActiveCfg = Debug|Win32
-		{6C82FDF0-ABCA-40FE-BE18-E900BC35F128}.Debug|x86.Build.0 = Debug|Win32
-		{6C82FDF0-ABCA-40FE-BE18-E900BC35F128}.Release|x64.ActiveCfg = Release|x64
-		{6C82FDF0-ABCA-40FE-BE18-E900BC35F128}.Release|x64.Build.0 = Release|x64
-		{6C82FDF0-ABCA-40FE-BE18-E900BC35F128}.Release|x86.ActiveCfg = Release|Win32
-		{6C82FDF0-ABCA-40FE-BE18-E900BC35F128}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {202FDEA2-EEDA-44FF-B825-79158D1E86B6}
-	EndGlobalSection
-EndGlobal
