--- image_contour/x64/Debug/image_contour.log
@@ -1,3 +1,0 @@
-﻿  img_contour.cpp
-C:\Users\KCCI25\source\opencv\opencv-0628\img_contour.cpp(56,28): warning C4267: '초기화 중': 'size_t'에서 'int'(으)로 변환하면서 데이터가 손실될 수 있습니다.
-  image_contour.vcxproj -> C:\Users\KCCI25\source\repos\image_contour\x64\Debug\image_contour.exe
