--- image_contour/x64/Debug/image_contour.tlog/image_contour.lastbuildstate
@@ -1,2 +1,0 @@
-PlatformToolSet=v142:VCToolArchitecture=Native32Bit:VCToolsVersion=14.29.30133:VCServicingVersionMFC=14.29.30136:VCServicingVersionCrtHeaders=14.29.30136:TargetPlatformVersion=10.0.19041.0:
-Debug|x64|C:\Users\KCCI25\source\repos\image_contour\|
