--- image_contour/x64/Debug/image_contour.vcxproj.FileListAbsolute.txt
@@ -1,1 +1,0 @@
-C:\Users\KCCI25\source\repos\image_contour\x64\Debug\image_contour.exe
